requirements.txt
matplotlib==3.2.2
numpy==1.20.2
opencv_python==4.5.4.60
pandas==1.3.4
Pillow==9.1.1
pydensecrf==1.0rc2
tensorboardX==2.5
torch==1.7.1
torchvision==0.8.0a0
